// src/core/http/generated/models/whatsAppTemplate.ts
/**
 * Generated by orval v8.14.0 🍺
 * Do not edit manually.
 * Petvex Admin API
 * OpenAPI spec version: 1.0.0
 */
import type { WhatsAppTemplateComponentsItem } from './whatsAppTemplateComponentsItem.ts';

export interface WhatsAppTemplate {
  key: string;
  name: string;
  meta_template_name: string;
  language: string;
  category: string;
  description?: string;
  components?: WhatsAppTemplateComponentsItem[];
  is_active: boolean;
}